fix: gcr table on manageNative
+ fix: forge error GCR...Routines.ts
+ return inserted account on createAccount
+ change balance column to bigint

+++ src/model/entities/GCRv2/GCR_Main.ts
@@ -11,7 +11,7 @@ export class GCR_Main {
     assignedTxs: string[]
     @Column({ type: "integer", name: "nonce" })
     nonce: number
-    @Column({ type: "integer", name: "balance" })
+    @Column({ type: "bigint", name: "balance" })
     balance: number
     @Column({ type: "jsonb", name: "identities" })
     identities: StoredIdentities
